requirements.txt


pandas==2.3.0
streamlit==1.45.1
google-genai==1.19.0
